Fixed a bug where forced metadata refresh wouldn't trigger the volume to update it's image.
