+++ Kavita.Common/KavitaException.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Runtime.Serialization;
 
 namespace Kavita.Common
 {
